refactor: reverted changes to poster dialog
