+++ apps/journeys-admin/src/components/Editor/VideoBlockEditor/Settings/Poster/Dialog/index.ts
@@ -1,0 +1,1 @@
+export { VideoBlockEditorSettingsPosterDialog } from './VideoBlockEditorSettingsPosterDialog'
